Fix: Создает папку mokky и переносит туда данные
