+++ src/const.js
@@ -37,47 +37,6 @@ export const TYPE_EVENTS = [
   },
 ];
 
-export const BUTTON_OFFERS = [
-  {
-    title: "Add luggage",
-    price: "50",
-    value: "luggage",
-    isChecked: true,
-  },
-  {
-    title: "Switch to comfort",
-    price: "80",
-    value: "comfort",
-    isChecked: true,
-  },
-  {
-    title: "Add meal",
-    price: "15",
-    value: "meal",
-    isChecked: false,
-  },
-  {
-    title: "Choose seats",
-    price: "5",
-    value: "meal",
-    isChecked: false,
-  },
-  {
-    title: "Travel by train",
-    price: "40",
-    value: "train",
-    isChecked: false,
-  },
-];
-
-export const DESTINATION_IMAGES = [
-  "img/photos/1.jpg",
-  "img/photos/2.jpg",
-  "img/photos/3.jpg",
-  "img/photos/4.jpg",
-  "img/photos/5.jpg",
-];
-
 export const SORT_VALUES = [
   {
     value: "Day",
